Refactor probably done
