+++ src/actionTypes/index.ts
@@ -1,38 +1,38 @@
-export const LOAD_INIT_DATA: string = 'LOAD_INIT_DATA';
-export const ADD_COMPONENT: string = 'ADD_COMPONENT';
 export const ADD_CHILD: string = 'ADD_CHILD';
-export const DELETE_CHILD: string = 'DELETE_CHILD';
-export const UPDATE_COMPONENT: string = 'UPDATE_COMPONENT';
-export const DELETE_COMPONENT: string = 'DELETE_COMPONENT';
-export const CHANGE_FOCUS_COMPONENT: string = 'CHANGE_FOCUS_COMPONENT';
+export const ADD_COMPONENT: string = 'ADD_COMPONENT';
+export const ADD_PROP: string = 'ADD_PROP';
 export const CHANGE_COMPONENT_FOCUS_CHILD: string = 'CHANGE_COMPONENT_FOCUS_CHILD';
 export const CHANGE_FOCUS_CHILD: string = 'CHANGE_FOCUS_CHILD';
-export const UPDATE_CHILDREN: string = 'UPDATE_CHILDREN';
-export const REASSIGN_PARENT: string = 'REASSIGN_PARENT';
-export const SET_SELECTABLE_PARENTS: string = 'SET_SELECTABLE_PARENTS';
+export const CHANGE_FOCUS_COMPONENT: string = 'CHANGE_FOCUS_COMPONENT';
+export const CHANGE_IMAGE_PATH: string = 'CHANGE_IMAGE_PATH';
+export const CHANGE_IMAGE_SOURCE: string = 'CHANGE_IMAGE_SOURCE';
+export const CHANGE_TUTORIAL: string = 'CHANGE_TUTORIAL';
+export const CREATE_APPLICATION: string = 'CREATE_APPLICATION';
+export const CREATE_APPLICATION_ERROR: string = 'CREATE_APPLICATION_ERROR';
+export const CREATE_APPLICATION_SUCCESS: string = 'CREATE_APPLICATION_SUCCESS';
+export const DELETE_ALL_DATA: string = 'DELETE_ALL_DATA';
+export const DELETE_CHILD: string = 'DELETE_CHILD';
+export const DELETE_COMPONENT: string = 'DELETE_COMPONENT';
+export const DELETE_IMAGE: string = 'DELETE_IMAGE';
+export const DELETE_PROP: string = 'DELETE_PROP';
+export const EDIT_COMPONENT: string = 'EDIT_COMPONENT';
+export const EDIT_MODE: string = 'EDIT_MODE';
 export const EXPORT_FILES: string = 'EXPORT_FILES';
-export const EXPORT_FILES_SUCCESS: string = 'EXPORT_FILES_SUCCESS';
 export const EXPORT_FILES_ERROR: string = 'EXPORT_FILES_ERROR';
+export const EXPORT_FILES_SUCCESS: string = 'EXPORT_FILES_SUCCESS';
 export const HANDLE_CLOSE: string = 'HANDLE_CLOSE';
 export const HANDLE_TRANSFORM: string = 'HANDLE_TRANSFORM';
-export const CREATE_APPLICATION: string = 'CREATE_APPLICATION';
-export const CREATE_APPLICATION_SUCCESS: string = 'CREATE_APPLICATION_SUCCESS';
-export const CREATE_APPLICATION_ERROR: string = 'CREATE_APPLICATION_ERROR';
+export const LOAD_INIT_DATA: string = 'LOAD_INIT_DATA';
 export const MOVE_TO_BOTTOM: string = 'MOVE_TO_BOTTOM';
 export const MOVE_TO_TOP: string = 'MOVE_TO_TOP';
 export const OPEN_EXPANSION_PANEL: string = 'OPEN_EXPANSION_PANEL';
-export const DELETE_PROP: string = 'DELETE_PROP';
-export const ADD_PROP: string = 'ADD_PROP';
-export const DELETE_ALL_DATA: string = 'DELETE_ALL_DATA';
-export const CHANGE_IMAGE_PATH: string = 'CHANGE_IMAGE_PATH';
-export const UPDATE_HTML_ATTR: string = 'UPDATE_HTML_ATTR';
-export const UPDATE_CHILDREN_SORT: string = 'UPDATE_CHILDREN_SORT';
-export const CHANGE_IMAGE_SOURCE: string = 'CHANGE_IMAGE_SOURCE';
-export const DELETE_IMAGE: string = 'DELETE_IMAGE';
-export const TOGGLE_STATE: string = 'TOGGLE_STATE';
-export const TOGGLE_CLASS: string = 'TOGGLE_CLASS';
-export const CHANGE_TUTORIAL: string = 'CHANGE_TUTORIAL';
-export const UNDO: string = 'UNDO';
+export const REASSIGN_PARENT: string = 'REASSIGN_PARENT';
 export const REDO: string = 'REDO';
-export const EDIT_MODE: string = 'EDIT_MODE';
-export const EDIT_COMPONENT: string = 'EDIT_COMPONENT';
+export const SET_SELECTABLE_PARENTS: string = 'SET_SELECTABLE_PARENTS';
+export const TOGGLE_CLASS: string = 'TOGGLE_CLASS';
+export const TOGGLE_STATE: string = 'TOGGLE_STATE';
+export const UNDO: string = 'UNDO';
+export const UPDATE_COMPONENT: string = 'UPDATE_COMPONENT';
+export const UPDATE_CHILDREN: string = 'UPDATE_CHILDREN';
+export const UPDATE_CHILDREN_SORT: string = 'UPDATE_CHILDREN_SORT';
+export const UPDATE_HTML_ATTR: string = 'UPDATE_HTML_ATTR';
