+++ src/actions/actionCreators.ts
@@ -7,70 +7,39 @@ import {
 } from '../intefaces/Interfaces';
 
 import {
-  LOAD_INIT_DATA,
-  ADD_COMPONENT,
   ADD_CHILD,
+  ADD_COMPONENT,
+  ADD_PROP,
+  CHANGE_COMPONENT_FOCUS_CHILD,
+  CHANGE_FOCUS_CHILD,
+  CHANGE_FOCUS_COMPONENT,
+  CHANGE_IMAGE_SOURCE,
+  CHANGE_TUTORIAL,
+  CREATE_APPLICATION,
+  CREATE_APPLICATION_ERROR,
+  CREATE_APPLICATION_SUCCESS,
+  DELETE_ALL_DATA,
   DELETE_CHILD,
   DELETE_COMPONENT,
-  TOGGLE_STATE,
-  TOGGLE_CLASS,
-  CHANGE_FOCUS_COMPONENT,
-  CHANGE_FOCUS_CHILD,
-  CHANGE_COMPONENT_FOCUS_CHILD,
+  DELETE_IMAGE,
+  DELETE_PROP,
+  EDIT_COMPONENT,
+  EDIT_MODE,
   EXPORT_FILES,
-  EXPORT_FILES_SUCCESS,
   EXPORT_FILES_ERROR,
+  EXPORT_FILES_SUCCESS,
   HANDLE_CLOSE,
   HANDLE_TRANSFORM,
-  CREATE_APPLICATION,
-  CREATE_APPLICATION_SUCCESS,
-  CREATE_APPLICATION_ERROR,
-  OPEN_EXPANSION_PANEL,
-  DELETE_PROP,
-  ADD_PROP,
-  DELETE_ALL_DATA,
-  UPDATE_HTML_ATTR,
-  CHANGE_IMAGE_SOURCE,
-  DELETE_IMAGE,
-  CHANGE_TUTORIAL,
-  UNDO,
+  LOAD_INIT_DATA,
   REDO,
-  EDIT_MODE,
-  EDIT_COMPONENT
+  TOGGLE_CLASS,
+  TOGGLE_STATE,
+  UNDO,
+  UPDATE_HTML_ATTR,
 } from '../actionTypes/index';
-
-import { loadState } from '../localStorage'; //this is a warning from 'localStorage' being a .js file instead of .ts. Convert to .ts to remove this warning.
+import { loadState } from '../localStorage';
 import createFiles from '../utils/createFiles.util';
 import createApplicationUtil from '../utils/createApplication.util';
-
-export const changeTutorial = (tutorial: number) => ({
-  type: CHANGE_TUTORIAL,
-  payload: { tutorial },
-});
-
-export const changeImagePath = (imageSource: string) => ({
-  type: CHANGE_IMAGE_SOURCE,
-  payload: { imageSource },
-});
-
-export const loadInitData = () => (dispatch: (arg: Action) => void) => {
-  loadState().then((data: any) => {
-    dispatch({
-      type: LOAD_INIT_DATA,
-      payload: {
-        data: data
-          ? { ...data.workspace, history: [], historyIndex: 0, future: [] } //erase history upon opening app
-          : {},
-      },
-    });
-  });
-};
-
-export const addComponent = ({ title }: { title: string }) => (
-  dispatch: (arg: Action) => void
-) => {
-  dispatch({ type: ADD_COMPONENT, payload: { title } });
-};
 
 export const addChild = ({
   title,
@@ -83,6 +52,27 @@ export const addChild = ({
 }) => (dispatch: (arg: Action) => void) => {
   dispatch({ type: ADD_CHILD, payload: { title, childType, HTMLInfo } });
 };
+
+export const addComponent = ({ title }: { title: string }) => (
+  dispatch: (arg: Action) => void
+) => {
+  dispatch({ type: ADD_COMPONENT, payload: { title } });
+};
+
+export const addProp = (prop: PropInt) => ({
+  type: ADD_PROP,
+  payload: { ...prop },
+});
+
+export const changeTutorial = (tutorial: number) => ({
+  type: CHANGE_TUTORIAL,
+  payload: { tutorial },
+});
+
+export const changeImagePath = (imageSource: string) => ({
+  type: CHANGE_IMAGE_SOURCE,
+  payload: { imageSource },
+});
 
 export const deleteChild = ({}) => (dispatch: (arg: Action) => void) => {
   // with no payload, it will delete focusd child
@@ -118,19 +108,6 @@ export const deleteComponent = ({
   dispatch({ type: DELETE_COMPONENT, payload: { componentId } });
 };
 
-export const changeFocusComponent = ({ title }: { title: string }) => (
-  dispatch: (arg: Action) => void
-) => {
-  dispatch({ type: CHANGE_FOCUS_COMPONENT, payload: { title } });
-};
-
-// make sure childId is being sent in
-export const changeFocusChild = ({ childId }: { childId: number }) => (
-  dispatch: (arg: Action) => void
-) => {
-  dispatch({ type: CHANGE_FOCUS_CHILD, payload: { childId } });
-};
-
 export const changeComponentFocusChild = ({
   componentId,
   childId,
@@ -144,9 +121,93 @@ export const changeComponentFocusChild = ({
   });
 };
 
+// make sure childId is being sent in
+export const changeFocusChild = ({ childId }: { childId: number }) => (
+  dispatch: (arg: Action) => void
+) => {
+  dispatch({ type: CHANGE_FOCUS_CHILD, payload: { childId } });
+};
+
+export const changeFocusComponent = ({ title }: { title: string }) => (
+  dispatch: (arg: Action) => void
+) => {
+  dispatch({ type: CHANGE_FOCUS_COMPONENT, payload: { title } });
+};
+
+export const createApplication = ({
+  path,
+  components = [],
+  genOption,
+  appName = 'reactype_app',
+  exportAppBool,
+}: {
+  path: string;
+  components: ComponentsInt;
+  genOption: number;
+  appName: string;
+  exportAppBool: boolean;
+}) => (dispatch: (arg: any) => void) => {
+  if (genOption === 0) {
+    exportAppBool = false;
+    dispatch(
+      exportFiles({
+        appName,
+        path,
+        components,
+        exportAppBool,
+      })
+    );
+  } else if (genOption) {
+    exportAppBool = true;
+    dispatch({
+      type: CREATE_APPLICATION,
+    });
+    createApplicationUtil({
+      path,
+      appName,
+      genOption,
+    })
+      .then(() => {
+        dispatch({
+          type: CREATE_APPLICATION_SUCCESS,
+        });
+        dispatch(
+          exportFiles({
+            appName,
+            path,
+            components,
+            exportAppBool,
+          })
+        );
+      })
+      .catch((err: string) =>
+        dispatch({
+          type: CREATE_APPLICATION_ERROR,
+          payload: { status: true, err },
+        })
+      );
+  }
+};
+
+export const deleteAllData = () => ({
+  type: DELETE_ALL_DATA,
+});
+
 export const deleteImage = () => ({
   type: DELETE_IMAGE,
 });
+
+export const deleteProp = (propId: number) => (
+  dispatch: (arg: Action) => void
+) => {
+  dispatch({ type: DELETE_PROP, payload: propId });
+};
+
+export const editComponent = ({ id, title }: { id: number; title: string }) => (
+  dispatch: (arg: Action) => void
+) => {
+  dispatch({ type: EDIT_COMPONENT, payload: { id, title } });
+};
 
 export const exportFiles = ({
   components,
@@ -204,75 +265,22 @@ export const handleTransform = (
   },
 });
 
-export const createApplication = ({
-  path,
-  components = [],
-  genOption,
-  appName = 'reactype_app',
-  exportAppBool,
-}: {
-  path: string;
-  components: ComponentsInt;
-  genOption: number;
-  appName: string;
-  exportAppBool: boolean;
-}) => (dispatch: (arg: any) => void) => {
-  if (genOption === 0) {
-    exportAppBool = false;
-    dispatch(
-      exportFiles({
-        appName,
-        path,
-        components,
-        exportAppBool,
-      })
-    );
-  } else if (genOption) {
-    exportAppBool = true;
+export const loadInitData = () => (dispatch: (arg: Action) => void) => {
+  loadState().then((data: any) => {
     dispatch({
-      type: CREATE_APPLICATION,
+      type: LOAD_INIT_DATA,
+      payload: {
+        data: data
+          ? { ...data.workspace, history: [], historyIndex: 0, future: [] } //erase history upon opening app
+          : {},
+      },
     });
-    createApplicationUtil({
-      path,
-      appName,
-      genOption,
-    })
-      .then(() => {
-        dispatch({
-          type: CREATE_APPLICATION_SUCCESS,
-        });
-        dispatch(
-          exportFiles({
-            appName,
-            path,
-            components,
-            exportAppBool,
-          })
-        );
-      })
-      .catch((err: string) =>
-        dispatch({
-          type: CREATE_APPLICATION_ERROR,
-          payload: { status: true, err },
-        })
-      );
-  }
+  });
 };
 
-export const openExpansionPanel = (component: ComponentInt) => ({
-  type: OPEN_EXPANSION_PANEL,
-  payload: { component },
+export const redo = () => ({
+  type: REDO,
 });
-
-export const deleteAllData = () => ({
-  type: DELETE_ALL_DATA,
-});
-
-export const deleteProp = (propId: number) => (
-  dispatch: (arg: Action) => void
-) => {
-  dispatch({ type: DELETE_PROP, payload: propId });
-};
 
 export const toggleComponentState = ({ id }: { id: number }) => (
   dispatch: (arg: Action) => void
@@ -286,32 +294,15 @@ export const toggleComponentClass = ({ id }: { id: number }) => (
   dispatch({ type: TOGGLE_CLASS, payload: { id } });
 };
 
-export const addProp = (prop: PropInt) => ({
-  type: ADD_PROP,
-  payload: { ...prop },
-});
-
-//action creators for undo and redo
-export const undo = () => ({
-  type: UNDO,
-});
-
-export const redo = () => ({
-  type: REDO,
-});
-
 export const toggleEditMode = ({ id }: { id: number }) => (
   dispatch: (arg: Action) => void
 ) => {
   dispatch({ type: EDIT_MODE, payload: { id } });
 };
-
-export const editComponent = ({ id, title }: { id: number, title: string }) => (
-  dispatch: (arg: Action) => void
-) => {
-  dispatch({ type: EDIT_COMPONENT , payload: { id, title } });
-};
- 
+//action creators for undo and redo
+export const undo = () => ({
+  type: UNDO,
+});
 
 export const updateHtmlAttr = ({
   attr,
